requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
b2sdk==2.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-environ==0.12.0
django-storages==1.14.5
idna==3.10
logfury==1.0.1
pillow==11.1.0
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.13.0
urllib3==2.3.0
